Added admin options

+++ backend/center/app/modules/users/controller.py
@@ -1,4 +1,4 @@
-from flask import abort, jsonify, request
+from flask import abort, jsonify, request, session
 from .model import User
 from werkzeug.security import generate_password_hash, check_password_hash
 
@@ -30,6 +30,8 @@ class UsersController:
 
             if res is None:
                 self.db.insert_one(user_dct)
+                user_dct.pop("_id")
+                user_dct.pop("password")
                 return jsonify(user_dct)
 
             return res if type(res) is not User else abort(400, "User already exists.")
@@ -40,10 +42,25 @@ class UsersController:
     def delete_user(self, data):
         try:
             email = data["email"]
-            self.db.delete_one({"email": email})
-            return email
+            if email == session["email"]:
+                return abort(
+                    400, "Cannot delete user in session (Cannot delete yourself)."
+                )
+            result = self.db.delete_one({"email": email})
+            if result.deleted_count == 1:
+                return jsonify(data)
         except KeyError:
             return abort(400, "Missing data to delete user.")
+        return abort(400, "No user found.")
+
+    def get_users(self):
+        query = self.db.find({})
+        users = [
+            {"email": user["email"]}
+            for user in query
+            if user["email"] != session["email"]
+        ]
+        return jsonify(users)
 
     def _get_user(self, data):
         try:
@@ -55,4 +72,3 @@ class UsersController:
             return None
         except KeyError:
             return abort(400, "Data does not have an email.")
-
